Implemented whisper/doc.go refactor
