+++ whisper/message_test.go
@@ -24,6 +24,9 @@ import (
 	"testing"
 	"time"
 
+	"github.com/stretchr/testify/require"
+
+	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/rlp"
@@ -492,4 +495,15 @@ func TestValidateAndParseSizeOfPayloadSize(t *testing.T) {
 			msg.ValidateAndParse()
 		})
 	}
+}
+
+func TestEncodeDecodeVersionedResponse(t *testing.T) {
+	response := NewMessagesResponse(common.Hash{1}, []EnvelopeError{{Code: 1}})
+	bytes, err := rlp.EncodeToBytes(response)
+	require.NoError(t, err)
+	var mresponse MultiVersionResponse
+	require.NoError(t, rlp.DecodeBytes(bytes, &mresponse))
+	v1resp, err := mresponse.DecodeResponse1()
+	require.NoError(t, err)
+	require.Equal(t, response.Response.Hash, v1resp.Hash)
 }
